service worker: Fix navigation-preload/resource-timing.https.html test
Before this CL, this test chains Response#text() and Response creation
inside the callback of FetchEvent#preloadResponse. This seems wrong as
preloadResponse can be resolved before reading response body. This CL
moves response creation out of the callback so that we can make sure
that preloadResponse will be always resolved after response body is
consumed, which is required to make the test work.

Bug: 872078
Change-Id: I6ecfcb8e0721df541d5a32f56ec4c062efbffdae