Implement ignore-opens-during-unload counter in Document
The removed test fast/loader/document-destruction-within-unload.html was
originally added in 8809405d796cf3023e26cefd4b06f369eb67f125, to make
sure that a document.write() call that blows away the current page still
works in unload event listeners and blocks any current attempt to
navigate. Since we no longer allow that to happen, the test is obsolete.
In fact, it can be seen as contradictory to
external/wpt/html/browsers/browsing-the-web/unloading-documents/005.html,
which tests that document.open() in an unload event doesn't block
navigation.

Bug: 583586, 866274
Change-Id: I99b35dd28c97e8603455805b31d49644bc7b23a5